+++ tests/unit/components/playback/controls/BarcodeViewer.spec.js
@@ -118,7 +118,7 @@ describe("BarcodeViewer.vue", () => {
     expect(wrapper.find("input").text()).toBe("");
   });
   test("Given that barcode entry is in manual mode, When a plate barcode is entered, Then it is stored in Vuex playback store correctly", async () => {
-    const spied_text_validator = jest.spyOn(TextValidation.prototype, "validate_plate_barcode");
+    const spied_text_validator = jest.spyOn(TextValidation.prototype, "validate_barcode");
 
     store.commit("flask/set_barcode_manual_mode", true);
     expect(store.state.playback.barcodes.plate_barcode.value).toBeNull();
@@ -136,7 +136,7 @@ describe("BarcodeViewer.vue", () => {
 
     await wrapper.vm.handle_manual_mode_choice(true);
     wrapper.find("input").setValue(test_barcode);
-    expect(spied_text_validator).toHaveBeenCalledWith(test_barcode, false);
+    expect(spied_text_validator).toHaveBeenCalledWith(test_barcode, "plate_barcode", false);
     expect(store.state.playback.barcodes.plate_barcode.value).toBe(test_barcode);
   });
   test("Given that its in manual mode and a valid barcode has been entered and playback state is BUFFERING, When Playback State is mutated to BUFFERING, Then the text of the Barcode Input remains as the valid barcode instead of becoming blank", async () => {