+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Added:
 ^^^^^^
 - Ability to start live view, recording, and stimulation from stim studio start button
 - Ability to select from previously used usernames when logging into the SettingsForm
+- Check for type of barcode enter into inputs
+- Added log when user set barcode manually
 
 Changed:
 ^^^^^^^^